+++ pkg/services/ngalert/api/generated_base_api_prometheus.go
@@ -1,17 +1,14 @@
-/*Package api contains base API implementation of unified alerting
+/* Package api contains base API implementation of unified alerting
  *
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  *
- * Need to remove unused imports.
+ * Do not manually edit these files, please find ngalert/api/swagger-codegen/ for commands on how to generate them.
  */
 package api
 
 import (
-	"net/http"
-
 	"github.com/grafana/grafana/pkg/api/response"
 	"github.com/grafana/grafana/pkg/api/routing"
-	"github.com/grafana/grafana/pkg/infra/log"
 	"github.com/grafana/grafana/pkg/middleware"
 	"github.com/grafana/grafana/pkg/models"
 )
@@ -21,25 +18,9 @@ type PrometheusApiService interface {
 	RouteGetRuleStatuses(*models.ReqContext) response.Response
 }
 
-type PrometheusApiBase struct {
-	log log.Logger
-}
-
 func (api *API) RegisterPrometheusApiEndpoints(srv PrometheusApiService) {
 	api.RouteRegister.Group("", func(group routing.RouteRegister) {
 		group.Get(toMacaronPath("/api/prometheus/{Recipient}/api/v1/alerts"), routing.Wrap(srv.RouteGetAlertStatuses))
 		group.Get(toMacaronPath("/api/prometheus/{Recipient}/api/v1/rules"), routing.Wrap(srv.RouteGetRuleStatuses))
 	}, middleware.ReqSignedIn)
-}
-
-func (base PrometheusApiBase) RouteGetAlertStatuses(c *models.ReqContext) response.Response {
-	recipient := c.Params(":Recipient")
-	base.log.Info("RouteGetAlertStatuses: ", "Recipient", recipient)
-	return response.Error(http.StatusNotImplemented, "", nil)
-}
-
-func (base PrometheusApiBase) RouteGetRuleStatuses(c *models.ReqContext) response.Response {
-	recipient := c.Params(":Recipient")
-	base.log.Info("RouteGetRuleStatuses: ", "Recipient", recipient)
-	return response.Error(http.StatusNotImplemented, "", nil)
 }
