+++ pkg/services/ngalert/api/legacy_api_routes.go
@@ -4,11 +4,12 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/grafana/grafana/pkg/expr/translate"
+
 	apimodels "github.com/grafana/alerting-api/pkg/api"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 	"github.com/grafana/grafana/pkg/api/response"
 	"github.com/grafana/grafana/pkg/bus"
-	"github.com/grafana/grafana/pkg/expr/translate"
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/services/ngalert/eval"
 	ngmodels "github.com/grafana/grafana/pkg/services/ngalert/models"
@@ -16,6 +17,252 @@ import (
 	"github.com/grafana/grafana/pkg/util"
 	"github.com/prometheus/common/model"
 )
+
+// listAlertInstancesEndpoint handles GET /api/alert-instances.
+func (api *API) listAlertInstancesEndpoint(c *models.ReqContext) response.Response {
+	cmd := ngmodels.ListAlertInstancesQuery{DefinitionOrgID: c.SignedInUser.OrgId}
+
+	if err := api.Store.ListAlertInstances(&cmd); err != nil {
+		return response.Error(500, "Failed to list alert instances", err)
+	}
+
+	return response.JSON(200, cmd.Result)
+}
+
+// conditionEvalEndpoint handles POST /api/alert-definitions/eval.
+func (api *API) conditionEvalEndpoint(c *models.ReqContext, cmd ngmodels.EvalAlertConditionCommand) response.Response {
+	evalCond := ngmodels.Condition{
+		Condition: cmd.Condition,
+		OrgID:     c.SignedInUser.OrgId,
+		Data:      cmd.Data,
+	}
+	if err := api.validateCondition(evalCond, c.SignedInUser, c.SkipCache); err != nil {
+		return response.Error(400, "invalid condition", err)
+	}
+
+	now := cmd.Now
+	if now.IsZero() {
+		now = timeNow()
+	}
+
+	evaluator := eval.Evaluator{Cfg: api.Cfg}
+	evalResults, err := evaluator.ConditionEval(&evalCond, timeNow(), api.DataService)
+	if err != nil {
+		return response.Error(400, "Failed to evaluate conditions", err)
+	}
+
+	frame := evalResults.AsDataFrame()
+
+	return response.JSONStreaming(200, util.DynMap{
+		"instances": []*data.Frame{&frame},
+	})
+}
+
+// alertDefinitionEvalEndpoint handles GET /api/alert-definitions/eval/:alertDefinitionUID.
+func (api *API) alertDefinitionEvalEndpoint(c *models.ReqContext) response.Response {
+	alertDefinitionUID := c.Params(":alertDefinitionUID")
+
+	condition, err := api.LoadAlertCondition(alertDefinitionUID, c.SignedInUser.OrgId)
+	if err != nil {
+		return response.Error(400, "Failed to load alert definition conditions", err)
+	}
+
+	if err := api.validateCondition(*condition, c.SignedInUser, c.SkipCache); err != nil {
+		return response.Error(400, "invalid condition", err)
+	}
+
+	evaluator := eval.Evaluator{Cfg: api.Cfg}
+	evalResults, err := evaluator.ConditionEval(condition, timeNow(), api.DataService)
+	if err != nil {
+		return response.Error(400, "Failed to evaluate alert", err)
+	}
+	frame := evalResults.AsDataFrame()
+
+	return response.JSONStreaming(200, util.DynMap{
+		"instances": []*data.Frame{&frame},
+	})
+}
+
+// getAlertDefinitionEndpoint handles GET /api/alert-definitions/:alertDefinitionUID.
+func (api *API) getAlertDefinitionEndpoint(c *models.ReqContext) response.Response {
+	alertDefinitionUID := c.Params(":alertDefinitionUID")
+
+	query := ngmodels.GetAlertDefinitionByUIDQuery{
+		UID:   alertDefinitionUID,
+		OrgID: c.SignedInUser.OrgId,
+	}
+
+	if err := api.Store.GetAlertDefinitionByUID(&query); err != nil {
+		return response.Error(500, "Failed to get alert definition", err)
+	}
+
+	return response.JSON(200, &query.Result)
+}
+
+// deleteAlertDefinitionEndpoint handles DELETE /api/alert-definitions/:alertDefinitionUID.
+func (api *API) deleteAlertDefinitionEndpoint(c *models.ReqContext) response.Response {
+	alertDefinitionUID := c.Params(":alertDefinitionUID")
+
+	cmd := ngmodels.DeleteAlertDefinitionByUIDCommand{
+		UID:   alertDefinitionUID,
+		OrgID: c.SignedInUser.OrgId,
+	}
+
+	if err := api.Store.DeleteAlertDefinitionByUID(&cmd); err != nil {
+		return response.Error(500, "Failed to delete alert definition", err)
+	}
+
+	return response.Success("Alert definition deleted")
+}
+
+// updateAlertDefinitionEndpoint handles PUT /api/alert-definitions/:alertDefinitionUID.
+func (api *API) updateAlertDefinitionEndpoint(c *models.ReqContext, cmd ngmodels.UpdateAlertDefinitionCommand) response.Response {
+	cmd.UID = c.Params(":alertDefinitionUID")
+	cmd.OrgID = c.SignedInUser.OrgId
+
+	evalCond := ngmodels.Condition{
+		Condition: cmd.Condition,
+		OrgID:     c.SignedInUser.OrgId,
+		Data:      cmd.Data,
+	}
+	if err := api.validateCondition(evalCond, c.SignedInUser, c.SkipCache); err != nil {
+		return response.Error(400, "invalid condition", err)
+	}
+
+	if err := api.Store.UpdateAlertDefinition(&cmd); err != nil {
+		return response.Error(500, "Failed to update alert definition", err)
+	}
+
+	return response.JSON(200, cmd.Result)
+}
+
+// createAlertDefinitionEndpoint handles POST /api/alert-definitions.
+func (api *API) createAlertDefinitionEndpoint(c *models.ReqContext, cmd ngmodels.SaveAlertDefinitionCommand) response.Response {
+	cmd.OrgID = c.SignedInUser.OrgId
+
+	evalCond := ngmodels.Condition{
+		Condition: cmd.Condition,
+		OrgID:     c.SignedInUser.OrgId,
+		Data:      cmd.Data,
+	}
+	if err := api.validateCondition(evalCond, c.SignedInUser, c.SkipCache); err != nil {
+		return response.Error(400, "invalid condition", err)
+	}
+
+	if err := api.Store.SaveAlertDefinition(&cmd); err != nil {
+		return response.Error(500, "Failed to create alert definition", err)
+	}
+
+	return response.JSON(200, cmd.Result)
+}
+
+// listAlertDefinitions handles GET /api/alert-definitions.
+func (api *API) listAlertDefinitions(c *models.ReqContext) response.Response {
+	query := ngmodels.ListAlertDefinitionsQuery{OrgID: c.SignedInUser.OrgId}
+
+	if err := api.Store.GetOrgAlertDefinitions(&query); err != nil {
+		return response.Error(500, "Failed to list alert definitions", err)
+	}
+
+	return response.JSON(200, util.DynMap{"results": query.Result})
+}
+
+func (api *API) pauseScheduler() response.Response {
+	err := api.Schedule.Pause()
+	if err != nil {
+		return response.Error(500, "Failed to pause scheduler", err)
+	}
+	return response.JSON(200, util.DynMap{"message": "alert definition scheduler paused"})
+}
+
+func (api *API) unpauseScheduler() response.Response {
+	err := api.Schedule.Unpause()
+	if err != nil {
+		return response.Error(500, "Failed to unpause scheduler", err)
+	}
+	return response.JSON(200, util.DynMap{"message": "alert definition scheduler unpaused"})
+}
+
+// alertDefinitionPauseEndpoint handles POST /api/alert-definitions/pause.
+func (api *API) alertDefinitionPauseEndpoint(c *models.ReqContext, cmd ngmodels.UpdateAlertDefinitionPausedCommand) response.Response {
+	cmd.OrgID = c.SignedInUser.OrgId
+	cmd.Paused = true
+
+	err := api.Store.UpdateAlertDefinitionPaused(&cmd)
+	if err != nil {
+		return response.Error(500, "Failed to pause alert definition", err)
+	}
+	return response.JSON(200, util.DynMap{"message": fmt.Sprintf("%d alert definitions paused", cmd.ResultCount)})
+}
+
+// alertDefinitionUnpauseEndpoint handles POST /api/alert-definitions/unpause.
+func (api *API) alertDefinitionUnpauseEndpoint(c *models.ReqContext, cmd ngmodels.UpdateAlertDefinitionPausedCommand) response.Response {
+	cmd.OrgID = c.SignedInUser.OrgId
+	cmd.Paused = false
+
+	err := api.Store.UpdateAlertDefinitionPaused(&cmd)
+	if err != nil {
+		return response.Error(500, "Failed to unpause alert definition", err)
+	}
+	return response.JSON(200, util.DynMap{"message": fmt.Sprintf("%d alert definitions unpaused", cmd.ResultCount)})
+}
+
+// LoadAlertCondition returns a Condition object for the given alertDefinitionID.
+func (api *API) LoadAlertCondition(alertDefinitionUID string, orgID int64) (*ngmodels.Condition, error) {
+	q := ngmodels.GetAlertDefinitionByUIDQuery{UID: alertDefinitionUID, OrgID: orgID}
+	if err := api.Store.GetAlertDefinitionByUID(&q); err != nil {
+		return nil, err
+	}
+	alertDefinition := q.Result
+
+	err := api.Store.ValidateAlertDefinition(alertDefinition, true)
+	if err != nil {
+		return nil, err
+	}
+
+	return &ngmodels.Condition{
+		Condition: alertDefinition.Condition,
+		OrgID:     alertDefinition.OrgID,
+		Data:      alertDefinition.Data,
+	}, nil
+}
+
+func (api *API) validateCondition(c ngmodels.Condition, user *models.SignedInUser, skipCache bool) error {
+	var refID string
+
+	if len(c.Data) == 0 {
+		return nil
+	}
+
+	for _, query := range c.Data {
+		if c.Condition == query.RefID {
+			refID = c.Condition
+		}
+
+		datasourceUID, err := query.GetDatasource()
+		if err != nil {
+			return err
+		}
+
+		isExpression, err := query.IsExpression()
+		if err != nil {
+			return err
+		}
+		if isExpression {
+			continue
+		}
+
+		_, err = api.DatasourceCache.GetDatasourceByUID(datasourceUID, user, skipCache)
+		if err != nil {
+			return fmt.Errorf("failed to get datasource: %s: %w", datasourceUID, err)
+		}
+	}
+
+	if refID == "" {
+		return fmt.Errorf("condition %s not found in any query or expression", c.Condition)
+	}
+	return nil
+}
 
 // conditionEvalEndpoint handles POST /api/alert-definitions/evalOld.
 func (api *API) conditionEvalOldEndpoint(c *models.ReqContext) response.Response {
@@ -332,4 +579,20 @@ func transExecErr(s string) (ngmodels.ExecutionErrorState, error) {
 		return ngmodels.KeepLastStateErrState, nil
 	}
 	return ngmodels.AlertingErrState, fmt.Errorf("unrecognized Execution Error setting %v", s)
+}
+
+func (api *API) validateOrgAlertDefinition(c *models.ReqContext) {
+	uid := c.ParamsEscape(":alertDefinitionUID")
+
+	if uid == "" {
+		c.JsonApiErr(403, "Permission denied", nil)
+		return
+	}
+
+	query := ngmodels.GetAlertDefinitionByUIDQuery{UID: uid, OrgID: c.SignedInUser.OrgId}
+
+	if err := api.Store.GetAlertDefinitionByUID(&query); err != nil {
+		c.JsonApiErr(404, "Alert definition not found", nil)
+		return
+	}
 }
